UserService/requirements.txt
absl-py==1.4.0
aniso8601==9.0.1
asgiref==3.7.2
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
distlib==0.3.6
Django==4.2.3
django-environ==0.10.0
django-extensions==3.2.3
django-rest==0.8.7
django-seed==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Faker==18.11.2
filelock==3.12.2
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphene==3.2.2
graphene-django==3.1.3
graphql-core==3.2.3
graphql-relay==3.2.0
grpcio==1.56.0
h5py==3.9.0
idna==3.4
keras==2.13.1rc0
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
mysqlclient==2.2.0
numpy==1.25.0
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pika==1.3.2
platformdirs==3.8.1
promise==2.3
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0rc1
tensorflow-estimator==2.13.0rc0
tensorflow-macos==2.13.0rc1
termcolor==2.3.0
text-unidecode==1.3
toposort==1.10
typing_extensions==4.6.3
urllib3==1.26.16
virtualenv==20.23.1
Werkzeug==2.3.6
wrapt==1.15.0
